VertexDeployment/requirements.txt
# Core dependencies
torch>=1.13.0
torchvision>=0.14.0
Pillow>=9.0.0
numpy>=1.20.0

# TorchServe and related
torchserve>=0.7.0
torch-model-archiver>=0.7.0
torch-workflow-archiver>=0.7.0

# Google Cloud
google-cloud-storage>=2.0.0
google-cloud-aiplatform>=1.16.0

# Utilities
tensorboard>=2.8.0
psutil>=5.9.0
requests>=2.27.0 